backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
gunicorn==23.0.0
python-dotenv==1.0.1
python-multipart==0.0.20
pydantic==2.10.3
ollama==0.4.4
langgraph==0.2.60
langchain-core==0.3.28
numpy==1.26.4
scikit-learn==1.5.2
